backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
python-dotenv==1.0.0
cors==1.0.1
httpx==0.25.2
cryptography==42.0.0
PyJWT==2.10.1